Add files via upload
This script updates 3 of the ISR functions so that they are compatible with 2022 versions of different required packages such as ggplot2. 

Note lines 62-65 have been adjusted to get annual rate ratios from Poisson/negative binomial models

  ##calculate angle per year exp(angle*scaleby)
  scaleby<-3600*24*365.25
  for (c in c("angle", "angle.left.CI", "angle.right.CI"))
    summary.dataset[[c]]<-exp(summary.dataset[[c]]*scaleby)

as the model uses posixCT objects it calculates a rate ratio per second. Adjust these according to your own needs. So perhaps take the exp away if not using a log-link, or multiply just by 24 if you want a daily rate ratio. 

Additionally, this also adds an extra parameter to the plot and master functions, namely ylim.max, which allows you to set the upper limit of the y-axis so that if you want to create multiple plots, you can have the same limits for all of them. 

Similarly, you can change these further as per your own project's needs.